+++ src/types/divider.ts
@@ -1,6 +1,8 @@
 import { z } from 'zod/v4'
 import { DirectionEnum } from './general'
 
-export const DividerPropsSchema = z.object({ direction: DirectionEnum.default('vertical') })
+export const DividerPropsSchema = z.object({ direction: DirectionEnum.nullable() })
 
-export type DividerProps = z.infer<typeof DividerPropsSchema>
+export type DividerProps = {
+	direction?: z.infer<typeof DirectionEnum>
+}